+++ Gateway/Request/RefundMotoMerchantAccountDataBuilder.php
@@ -1,0 +1,36 @@
+<?php
+/**
+ *
+ * Adyen Payment module (https://www.adyen.com/)
+ *
+ * Copyright (c) 2022 Adyen N.V. (https://www.adyen.com/)
+ * See LICENSE.txt for license details.
+ *
+ * Author: Adyen <magento@adyen.com>
+ */
+
+namespace Adyen\Payment\Gateway\Request;
+
+use Adyen\Payment\Helper\ChargedCurrency;
+use Magento\Payment\Gateway\Request\BuilderInterface;
+
+/**
+ * Class MotoRefundDataBuilder
+ */
+class RefundMotoMerchantAccountDataBuilder implements BuilderInterface
+{
+    /**
+     * @param array $buildSubject
+     * @return array
+     */
+    public function build(array $buildSubject)
+    {
+        /** @var \Magento\Payment\Gateway\Data\PaymentDataObject $paymentDataObject */
+        $paymentDataObject = \Magento\Payment\Gateway\Helper\SubjectReader::readPayment($buildSubject);
+        $merchantAccount = $paymentDataObject->getPayment()->getAdditionalInformation('motoMerchantAccount');
+        $requestBody[] = ["merchantAccount" => $merchantAccount];
+
+        $request['body'] = $requestBody;
+        return $request;
+    }
+}
